Changed base image for worker to use Eclipse openjdk image.
The prior openjdk image has been deprecated and will no longer work
as of October 2022. So upgrading was needed. This image is not currently
working for me, in that the IndexWorker throws an exception on startup
due to a SOLR initiation issue. I am not sure if that problem was already
there in the branch.

In addition, I updated the image to make it easier to debug the image. If
idxworked.debug="TRUE" in values.yaml, then the new entrypoint.sh will start
an infinite loop rather than the java class.